Add metadata for logback-classic 1.2.11
This commit adds the minimal amount of metadata for logback-classic,
involving programmatic setups of the `LoggerContext`. This means that
this does not optimize for logback setups relying on `logback.xml`
configuration files.

Here, the `PatternLayout` is registering default converters coming from
both logback-classic and logback-core. The logback `Compiler`
reflectively instantiates those as they are needed by the configured
logging pattern layout.

This change contributes tests that look up all converters available by
default and ensures that they are interpreted as expected and that
parser errors do not happen at runtime.