Remove Log4j2 entries from netty-common's reflection metadata
Apache Log4j2 does not work in a native image and that will be the
case until Log4j2 3.0 at the earliest [1]. Furthermore, Netty's
InternalLoggerFactory is written such that SLF4J is preferred. As
such, it does not make sense for Netty to provide reflection
metadata for Log4j2 – in must cases the code that tries to use
Log4j2 won't be reached and when it is, Log4j2 will not be usable
anyway.

In addition to the reflection metadata for Log4j2 not being
beneficial to Netty, it's also harmful when using JBoss Logging. The
metadata is sufficient for JBoss Logging's attempt to use Log4j2 to
reach the point where error-level logging of stack traces is
performed before falling back to another logging framework. Without
Netty's metadata for Log4j2, this fallback is performed silently as
intended.

[1] https://issues.apache.org/jira/browse/LOG4J2-2649