Add metadata for Eclipse Angus Jakarta Mail reference implementation
As part of Jakarta EE 10, the coordinates of the Jakarta Mail
reference implementation have been updated with the group ID changing
from com.sun.mail to org.eclipse.angus. This makes the existing
metadata for com.sun.mail:jakarta.mail ineffective. This commit adds
metadata for the the new artifact, focusing on SMTP as the existing
metadata does.