Remove MBean-related entries from hibernate-core's metadata
Hibernate's metadata contains entries for various MBean-related
classes. This metadata causes a call to
ManagementFactory.getPlatformMBeanServer() to fail with a
javax.management.openmbean.OpenDataException. It would appear that
the metadata is sufficient for the bootstrapping of the MBean server
to take a different code path to normal and this path fails. Removing
the metadata fixes  ManagementFactory.getPlatformMBeanServer() and
has no adverse effects on the existing tests for hibernate-core.