Add Spring libraries
This commit adds the Spring Boot starters that are known to be
compatible with GraalVM Native and are tested in the Spring CI with
dedicated smoke tests.

Instead of listing all Spring artifacts that are known to work, we are
only listing the Spring Boot starters and optional artifacts that
developers add to their Spring Boot projects.

Spring Cloud projects are not listed in this commit as they are not
released yet on Maven Central.

Closes gh-146