Fix "requires" clauses being ignored
This commit fixes the metadata repository not being used in the TCK.
Before, it was assumed that each test was working in isolation, but
in practice, #211 showed that it is not the case.

So instead of the TCK figuring out by itself which metadata config
to include, and completely ignoring the "requires" clauses, we
will now simply configure the tests to use the metadata from the
current state of the repository. It makes things easier, at the
cost of possible side effects: it is possible that the metadata
from a transitive dependency is actually what makes a build pass.

However this drawback should be covered by regular code reviews,
as it is unexpected that metadata for a transitive "talks" about
the current module under test.

Fixes #211