Correct type reachable condition in Quartz resource config
fe74d0e8 added a type reachable condition for Quartz's SQL resource
files but the type that was used (org.quartz.SimpleJob) isn't part
of Quartz and only exists in the tests for the Quartz metadata.

This commit updates the tests to reveal the problem by moving their
classes out of the org.quartz namespace and updates the metadata to
use a type in the condition that's part of Quartz.