Add support for Netty 4.1.115+
This commit adds additional reflection metadata for MpscUnpaddedArrayQueueConsumerIndexField,
MpscUnpaddedArrayQueueProducerIndexField and MpscUnpaddedArrayQueueProducerLimitField
in the io.netty.util.internal.shaded.org.jctools.queues.unpadded package.

It also refines the metadata with a "default-for" entry that ensures that previous metadata
continue to be used for Netty 4.1.0 to 4.1.114, uses consistently Netty 4.1.80 for the tests
of the previous version, and adds related tests for Netty 4.1.115.