Add metadata for upcoming netty-common 4.2.3 build time flag https://github.com/netty/netty/pull/15325 introduced code that works around a bug in GraalVM 19-25 where native image would fail with a linker error. The fix requires the class CleanerJava25 to be initialized at build time, but the native-image.properties to do this in the netty repo is overridden by graalvm-reachability-metadata.
This PR adds the same flag to this repo for the upcoming 4.2.3 release.

Because the flag is actually counterproductive on the current netty release – with it, the build will fail even on GraalVM versions like 17 that don't have the bug in the first place – I can't add it to the existing metadata, and need to create a new version instead. I also can't test this change locally.

One thing I'm concerned about is whether native-image.properties works at all in this repo, since this is the first time it's used here.