Fix flyway-database-postgresql's transactional lock setting
The metadata for flyway-database-postgresql did not allow
org.flywaydb.database.postgresql.TransactionalModel to be serialized
to JSON. This resulted in a failure if the setTransactionalLock
method on PostgreSQLConfigurationExtension was called in a native
image.

This commit adds the necessary metadata and updates the existing
test to cover this code path.